UPSTREAM: cpuidle: poll_state: Revise loop termination condition
If need_resched() returns "false", breaking out of the loop in
poll_idle() will cause a new idle state to be selected, so in fact
it usually doesn't make sense to spin in it longer than the target
residency of the second state.  [Note that the "polling" state is
used only if there is at least one "real" state defined in addition
to it, so the second state is always there.]  On the other hand,
breaking out of it early (say in case the next state is disabled)
shouldn't hurt as it is polling anyway.

For this reason, make the loop in poll_idle() break if the CPU has
been spinning longer than the target residency of the second state
(the "polling" state can only be state[0]).

Change-Id: I0b598e6d0efa5eacf23ab3a0d5759e11a01cf994
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>