UPSTREAM: cpuidle: Introduce cpuidle_driver_state_disabled() for driver quirks
Commit 99e98d3fb100 ("cpuidle: Consolidate disabled state checks")
overlooked the fact that the imx6q and tegra20 cpuidle drivers use
the "disabled" field in struct cpuidle_state for quirks which trigger
after the initialization of cpuidle, so reading the initial value of
that field is not sufficient for those drivers.

In order to allow them to implement the quirks without using the
"disabled" field in struct cpuidle_state, introduce a new helper
function and modify them to use it.

Fixes: 99e98d3fb100 ("cpuidle: Consolidate disabled state checks")
Reported-by: Len Brown <lenb@kernel.org>
Change-Id: I418efffe962a74672d97ab30067c455658cefc8e
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>