arm64: Disable GENERIC_IRQ_EFFECTIVE_AFF_MASK
The effective affinity mask causes a lot of bugs by virtue of many
set_irq_affinity handlers only setting an effective affinity mask for an
IRQ's parent but not the IRQ itself. Since this is a widespread issue that
would require manual fixing on every different SoC, just disable the
effective affinity mask altogether and use the first CPU in an affinity
mask configured.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Dark-Matter7232 <me@const.eu.org>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Change-Id: Ia3c021b7db60d6efe59c463ca34441fb41cf0655
Signed-off-by: Nishant Kumar <www.rajsonu13@gmail.com>